requirements.txt
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangosaml2==1.9.3
djangosaml2idp==0.7.2
elementpath==4.4.0
idna==3.7
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pysaml2==7.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
xmlschema==2.5.1
